ext::Scan now handles all of uint/int/float. Wrapping up ext::Scan
