+++ examples_tests/51.RadixSort/CMakeLists.txt
@@ -3,8 +3,4 @@ if(NOT RES)
 	message(FATAL_ERROR "common.cmake not found. Should be in {repo_root}/cmake directory")
 endif()
 
-set(EXAMPLE_SOURCES
-	../../src/nbl/ext/Scan/Scan.cpp
-)
-
-nbl_create_executable_project("${EXAMPLE_SOURCES}" "" "" "")
+nbl_create_executable_project("" "" "" "")
