--- examples_tests/51.RadixSort/DummyDownsweepClient.comp
@@ -1,5 +1,0 @@
-#define _NBL_GLSL_EXT_SCAN_BIN_OP_ (1 << 6)
-
-layout (local_size_x = _NBL_GLSL_WORKGROUP_SIZE_) in;
-
-#include "nbl/builtin/glsl/ext/Scan/default_downsweep.comp"
