+++ include/nbl/builtin/glsl/ext/Scan/default_downsweep.comp
@@ -12,7 +12,8 @@ void main()
 {
 	uint idx = min(nbl_glsl_ext_Scan_Parameters_t_getElementCountTotal() - 1u,
 		(gl_GlobalInvocationID.x + 1u) * nbl_glsl_ext_Scan_Parameters_t_getStride() - 1u);
-	uint result = _NBL_GLSL_EXT_SCAN_DOWNSWEEP_TYPE_(idx);
+	nbl_glsl_ext_Scan_Storage_t result;
+	_NBL_GLSL_EXT_SCAN_DOWNSWEEP_TYPE_(idx, result);
 	nbl_glsl_ext_Scan_setData(idx, result);
 }
