+++ include/nbl/ext/Scan/Scan.h
@@ -12,6 +12,7 @@ namespace Scan
 typedef uint32_t uint;
 #include "nbl/builtin/glsl/ext/Scan/parameters_struct.glsl"
 
+template <typename T>
 class Scan final : public core::IReferenceCounted
 {
 public:
@@ -38,7 +39,27 @@ public:
 		std::stack<uint32_t> element_count_pass_stack;
 	};
 
-	Scan(video::IDriver* driver, Operator op, const uint32_t wg_size);
+	Scan(video::IDriver* driver, Operator op, const uint32_t wg_size) : m_wg_size(wg_size)
+	{
+		assert(nbl::core::isPoT(wg_size));
+
+		const asset::SPushConstantRange pc_range = { asset::ISpecializedShader::ESS_COMPUTE, 0u, sizeof(nbl_glsl_ext_Scan_Parameters_t) };
+		video::IGPUDescriptorSetLayout::SBinding binding = { 0u, asset::EDT_STORAGE_BUFFER, 1u, video::IGPUSpecializedShader::ESS_COMPUTE, nullptr };
+
+		m_ds_layout = driver->createGPUDescriptorSetLayout(&binding, &binding + 1);
+		m_pipeline_layout = driver->createGPUPipelineLayout(&pc_range, &pc_range + 1, core::smart_refctd_ptr(m_ds_layout));
+
+		const char* data_type_name;
+		if constexpr (std::is_same_v<T, uint32_t>)
+			data_type_name = "uint";
+		else if constexpr (std::is_same_v<T, int>)
+			data_type_name = "int";
+		else if constexpr (std::is_same_v<T, float>)
+			data_type_name = "float";
+
+		m_upsweep_pipeline = createPipeline("nbl/builtin/glsl/ext/Scan/default_upsweep.comp", op, data_type_name, driver);
+		m_downsweep_pipeline = createPipeline("nbl/builtin/glsl/ext/Scan/default_downsweep.comp", op, data_type_name, driver);
+	}
 
 	inline auto getDefaultDescriptorSetLayout() const { return m_ds_layout.get(); }
 
@@ -107,7 +128,16 @@ public:
 		video::COpenGLExtensionHandler::extGlMemoryBarrier(GL_SHADER_STORAGE_BARRIER_BIT);
 	}
 
-	static void updateDescriptorSet(video::IGPUDescriptorSet* set, core::smart_refctd_ptr<video::IGPUBuffer> descriptor, video::IVideoDriver* driver);
+	static inline void updateDescriptorSet(video::IGPUDescriptorSet* set, core::smart_refctd_ptr<video::IGPUBuffer> descriptor, video::IVideoDriver* driver)
+	{
+		video::IGPUDescriptorSet::SDescriptorInfo ds_info = {};
+		ds_info.desc = descriptor;
+		ds_info.buffer = { 0u, descriptor->getSize() };
+
+		video::IGPUDescriptorSet::SWriteDescriptorSet writes = { set, 0, 0u, 1u, asset::EDT_STORAGE_BUFFER, &ds_info };
+
+		driver->updateDescriptorSets(1, &writes, 0u, nullptr);
+	}
 
 	const uint32_t m_wg_size;
 
@@ -119,7 +149,37 @@ private:
 	core::smart_refctd_ptr<video::IGPUComputePipeline> m_upsweep_pipeline = nullptr;
 	core::smart_refctd_ptr<video::IGPUComputePipeline> m_downsweep_pipeline = nullptr;
 
-	core::smart_refctd_ptr<video::IGPUComputePipeline> createPipeline(const char* shader_include_name, Operator bin_op, video::IDriver* driver);
+	core::smart_refctd_ptr<video::IGPUComputePipeline> createPipeline(const char* shader_include_name, Operator bin_op, const char* data_type_name,
+		video::IDriver* driver)
+	{
+		const char* source_fmt =
+R"===(#version 430 core
+
+#define _NBL_GLSL_WORKGROUP_SIZE_ %u
+#define _NBL_GLSL_EXT_SCAN_BIN_OP_ %u
+#define _NBL_GLSL_EXT_SCAN_STORAGE_TYPE_ %s
+
+layout (local_size_x = _NBL_GLSL_WORKGROUP_SIZE_) in;
+ 
+#include "%s"
+
+)===";
+
+		// Todo: This just the value I took from FFT example, don't know how it is being computed.
+		const size_t extraSize = 4u + 8u + 8u + 128u;
+
+		auto shader = core::make_smart_refctd_ptr<asset::ICPUBuffer>(strlen(source_fmt) + extraSize + 1u);
+		snprintf(reinterpret_cast<char*>(shader->getPointer()), shader->getSize(), source_fmt, m_wg_size, bin_op, data_type_name, shader_include_name);
+
+		auto cpu_specialized_shader = core::make_smart_refctd_ptr<asset::ICPUSpecializedShader>(
+			core::make_smart_refctd_ptr<asset::ICPUShader>(std::move(shader), asset::ICPUShader::buffer_contains_glsl),
+			asset::ISpecializedShader::SInfo{ nullptr, nullptr, "main", asset::ISpecializedShader::ESS_COMPUTE });
+
+		auto gpu_shader = driver->createGPUShader(core::smart_refctd_ptr<const asset::ICPUShader>(cpu_specialized_shader->getUnspecialized()));
+		auto gpu_shader_specialized = driver->createGPUSpecializedShader(gpu_shader.get(), cpu_specialized_shader->getSpecializationInfo());
+
+		return driver->createGPUComputePipeline(nullptr, core::smart_refctd_ptr(m_pipeline_layout), std::move(gpu_shader_specialized));
+	}
 };
 
 }
